fixing bug + improve + rearrange
setting death to agent (there was an error when we select a unit and try to move it after he died).
Delete some uncessary stuff in code.
Setting team 1 to face left side sprite on awake (for the demonstration).
Setting selectionCircle on all prefabs/variants and set range+damage to all.
Adding attack animation to Knight
+ Horesman