Fix bugs crossbow and more
Crossbow arraw is destroyed afrter some time if didnt reached target destination (target moved).
When the arrow collides a world object that is not a target (like mountain/water etc.) the arrow is distroyed.
+ change the speed arrow to 150 (high speed)

Add step sound for the regular units (not horses) - the function called twice from the animation played.

Add UI canvas options on top of the screen in WarScene, and add UI control sound in WarScene