Implement Minimap
There is a mini-map that move according to the camera with option to click/drag inside the minimap to see other places in the map, so when click/drag in the minimap the main camera follow that mouse position so give a preview in the whole screen.
The units are displayed in the minimap as circles according to their team color.

Need to set boundries of the main camera at war scene like in the preparation phase so the camera/minimap don't show out of the map.