Create paths of training UI:
A "diagram" with the units and the buildings with all the paths possible for training the units through the different buildings.

The player can scroll up and down using the bar on the side or click and drag up/down with the mouse to see the whole paths.

Things to add:
- Button in Menu scene, in building and in training scene that when clicked will display the paths training canvas.
- Popup on each building that shows the cost for training the specific unit that will pass through this building.